feat: add more info to json output
this is in preparation for the eventual vim plugin that helps you
visualize things better. i have found that generating an HTML report is
pretty complex, and the experience might not be the most optimal for
everyone. with editor plugins, hopefully the experience in navigating
and viewing changes might be... enhanced.